requirements.txt
numpy
onvif_zeep
opencv-contrib-python-headless
requests
scikit-learn
